GPIO input problem was solved
PWM, input and output are working properly

+++ ATmega Firmata/1sheelds_functions.cpp
@@ -139,7 +139,7 @@ void analogWrite(uint8 pin, int val)
 	{
 		timer = digitalPinToTimer(pin);
 		pwm_Setup(timer);
-		pwm_SetDutyCycle(val, timer);
+		pwm_SetDutyCycle((uint8)val, timer);
 
 	}
